more speed

+++ inc/cleri/kwcache.h
@@ -13,10 +13,8 @@ typedef struct cleri_parse_s cleri_parse_t;
 
 /* private functions */
 uint8_t * cleri__kwcache_new(const char * str);
-ssize_t cleri__kwcache_match(cleri_parse_t * pr, const char * str);
+uint8_t cleri__kwcache_match(cleri_parse_t * pr, const char * str);
 void cleri__kwcache_free(uint8_t * kwcache);
-
-
 
 #endif /* CLERI_KWCACHE_H_ */
